requirements.txt
flask==2.2.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pycryptodome==3.19.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pyotp==2.9.0
requests==2.31.0
flask-limiter==3.5.0
Werkzeug==2.2.3
google-generativeai
datetime
google.genai
razorpay
supabase